requirements.txt
gdown
timm
huggingface_hub
transformers
pillow
